fix logic to save and display boxids when user changes boxids ready to be shipped
